+++ lowlink.hpp
@@ -3,7 +3,7 @@
 #include "graph_base.hpp"
 /**
  * @brief 関節点・橋を求めるアルゴリズム
- * 
+ *
  */
 struct lowlink {
 	const UnweightedGraph &g;
@@ -22,18 +22,20 @@ struct lowlink {
 		int visit_count = 0;
 		// label
 		label[at] = count++;
-		id[at]    = label[at];//id.labelは通常はidのほうが小さくなる、子はlabalが大きくなるはず
-		//深さ優先探索
+		id[at]    = label
+		    [at];  // id.labelは通常はidのほうが小さくなる、子はlabalが大きくなるはず
+		// 深さ優先探索
 		for (auto &itr : g[at]) {
-			if (itr == last && !first_last) {//親に初めて訪れる(二重辺防止)
+			if (itr == last && !first_last) {  // 親に初めて訪れる(二重辺防止)
 				first_last = true;
 
 			} else if (label[itr] == -1) {  // 初めて訪れる場合
 				// trace(at, "->", itr);
 
 				count = dfs(itr, at, count);
-				++visit_count;//訪問順の調整
-				chmin(id[at], id[itr]);  // apply id(子が逆辺を通ってたどりつける辺)
+				++visit_count;  // 訪問順の調整
+				chmin(id[at],
+				      id[itr]);  // apply id(子が逆辺を通ってたどりつける辺)
 				// det 関節点の検出(始点以外)
 				articulat |= (last != -1) && (id[itr] >= label[at]);
 
@@ -44,16 +46,14 @@ struct lowlink {
 					bridge.emplace_back(at, itr);
 				}
 
-			} else {//逆辺をたどって到達できる(1回のみ)
+			} else {  // 逆辺をたどって到達できる(1回のみ)
 				chmin(id[at], label[itr]);
 			}
 		}
 		// def 関節点の検出(始点)
 		articulat |= (last == -1) && (visit_count > 1);
-		if (articulat) {
-			articulation.push_back(at);
-			// add articlat
-		}
+		if (articulat) articulation.push_back(at);
+		// add articlat
 
 		return count;
 	}
